veyron/services/wspr: Updating WSPR and and JS API to work with the recent changes on separating application and system error
We still assume the last out arg is an error in wspr which is enforced
by IDL. JS services are always assumed to return an error.

If we allow services not to return error, we would need to change the code
in wspr.go which assumes last argument is an error to be more flexible
For instance it could inspect the types of out args from the IDL and
see if one is an error or not. Or if it sees more than two error out
args it can combine them. Right now error is anydata in IDL so inspection
is not possible.

Change-Id: Ifb80c79e30e11415b73778c19b667990c5f2d6b6