javascript/api: Switching all tests and sample code to use on('data') instead of on('readable')
Based on documentation(see ref), on('data') is the best way if one needs to read
the stream as fast as possible. There are also gotchas associated with
on('readable'), for instance stream.read() inside a 'readable' is not
gauranteed to only return one item and can return everything in the buffer
meaning calls to stream.read() need to be in a loop like:
while (null !== (chunk = readable.read())) {...} which we were not doing and code does not neccessarily break if no buffering is happening.

Using on('data') seems cleaner and behaviour is more consistent as it will
always return one item as it becomes available.

Change-Id: I055b63393848136183e436fb8241f006de163e3e
ref: http://nodejs.org/api/stream.html