veyron/services/wsprd: Changes to identity handling.
WSPR was expecting identities to be passed in from the web app with
every request, and it was caching clients based on the identity.  That
logic is all gone now.

In it's place is a new method "assocIdentity" which the extension will
call to associate an identity with a name and origin.  The identity is
stored in the identity manager.

There were other refactorings involved too.

Change-Id: Ic845bff46b60c08d64b7db66239068bb6b0c92f0