Adding a commit that was missed when veyron.js was forked from veyron/javascript/api.
For some reason https://veyron.googlesource.com/veyron/+/2dcb2742bb8dc15f181d7c665ff3ac4a8dd18771 is missing in Veyron.JS, adding it back.

Details of original CL:

javascript/api: Stubbing out the signature() method on the bound service
object in the JavaScript API.

This is needed as external code may need to know the service signature
for an object name.
Exposing signature it as regular method on the stub seems to make the
most sense as it is a callable method on the object and Go API already
stubs out signature() on the generated code too.


Change-Id: I759a37ceaa9806b4dabd3a067577e2aaaa738c76