Move various Grunt tasks to Make:   * make clean   * make lint   * make docs   * make build   * make test (calls ./vgrunt test)
The only stuff still left in Grunt is the test tasks.

The default make task is currently "make build".

The test artifacts have all moved to test_out/.  The dist/ directory is
only for the built and minified veyron.js files and the source map.

There are no more browser_only files.  Browserify takes care of that for
us.  This removes an extra level of indirection that was not needed.

Change-Id: I377dbf88e6579abbde794617a039c149fe82c23e