Move the fortunejs example into the veyron-www repo, and clean it up.
There is now just one client.js file and one server.js file.  We use
browserify to build the browser bundles, and the 'static' http server to
serve them.

The README has been updated to walk the user through all the steps.

There was one change needed in veyron.js, namely that the it needs to
re-try 'auth' requests until the extension reports that it received it.
The reason for this is that the web-app can start and request an auth
before the extension content script has started, so it needs to retry.

Change-Id: I18bb081d1364ce18e428fcec93131c47e578f5e2