veyron2/context: Allow the runtime to be obtained from a context.T
This was a little uglier than I thought it would be.  The problem is with
code that handles contexts inside the runtime.  When you write tests for
parts of the runtime implementation you can't really have a runtime so
you can't really have contexts with a reference.  You have to pass nil.

Change-Id: Ia25180e9dee2b8d862445ff7ddf67810c8663f8f