Adds temporary fix for debugging promise wrapped exceptions.
Exceptions within blocks guarded by the es6-promise library disappear when
the promise.catch(fn) invocation is omitted. Furthermore errors within the
scope of the catch callback cannot be thrown (converted back to an
exception).

This change adds a quick fix that breaks the offending error out of the
promise call-stack so that exceptions can be discovered in an immediate and
natural way.

Change-Id: Id617d7423057d1ce6cf562fd5147f06d111ded69