Adds new integration test runner.
This is an initial pass at the integration test runner that will replace the
grunt/vgrunt setup we currently have. As of this CL those pieces are still
in-place and will be removed in subsequent CLs.

This is effectively re-implementation of what resides in vgrunt and
grunt_tasks/integration_test_environment_setup.js. The pieces have been
broken up into modules that handle setup and running of dependent services
which can be re-used later by other veyron JS
applications.

The runner is a wrapper around prova and will eventually pass it's CLI
arguments through. This will allow an easy way to define new types of tests
that rely on different sets of services or on having services spin up for
each test (like for storage tests).

Change-Id: I944f0ae4f51264c2ab90b00987b7d6ec41339ad1