Allow users to create a runtime with an unauthenticated id, by setting "config.requireAuthenticatedId = false".
This is useful for applications that don't rely on an authenticated
identity, and for developing applications when you don't want to depend
on a connection to the identity server.

This CL also adds a timeout to the getIdentity call.  It responds with
an error if veyron.js cannot commounicate with the extension for 5
seconds.

Fixes https://code.google.com/p/envyor/issues/detail?id=315

Change-Id: I1776a6677fdc1af67aa9819f23dd1e2c30a0264c