veyron.js: Moving wspr_sample into veyron.js repo and having veyron.js build all the binaries it needs including wsprd, proxyd, etc... itself.
-wspr_sample which is used solely for integration tests is moved out of
roadmap.examples and into veyron.js repo. History is preserved.

-I am also changing the veyron.js build process so in addition to building
wspr_sample, it also builds everything it needs (mounttabled, wsprd, etc...)
in its own bin folder. This eliminates the need for the person running the tests
to also install veyron binaries which used to cause flaky behaviour when
binaries were out-of-date. The build target only rebuilds if a source file in
wspr_sample or in veyron.io changes.

Change-Id: I8c1524e1e0d3c51732ef0ead429110adf1b0b506