"veyron.js": New Security Model
This CL replaces replaces all old security model constructs in the
JavaScript code to new ones. Below are the major changes:
1) The type PublicId is replaced with Blessings, and PrivateId with Principal.

2) Blessings does not have a 'names' field (which the old PublicId had)
nor does it support the ForContext method at present (which the Go
implementation does). However, the blessing names of the remote end of an IPC
are exposed via the "remoteBlessingStrings" field of the (security) context.
This field is exclusive to the JS definition of context and is not present in
the Go or Java definitions.

3) Three new WSPR message_types are addedd for blessing public keys
(BLESS_PUBLICKEY), unlinking blessings (UNLINK_BLESSINGS), and creating
new random self-signed blessings (NEW_BLESSINGS)

Change-Id: I79fb76b9d41af6b45f6afc52b307ca7fa73c3fe0