veyron.js: fixes flakiness with FlushCacheEntry test.
After some rather insane concurrency debugging, the issue
was discovered as the following:
Between the ns.resolve() (which is supposed to cache the name) and
ns.flushCacheEntry() (which is supposed to remove the cache entry and return
true if anything was flushed) there is a little time where, based on
logs I saw, the server responsible for /house/alarm can remount itself.
Now based on the logs, when it does remount, is uses a different
crypto identifier as part of its endpoint which means, as far as cache
is concerned, it was never there, since it's a different endpoint.

I am changing the test to just verify that FlushCacheEntry goes
through, ignoring whether anything was actually flushed on not.

Change-Id: I8427888ed12f487604e52a9af287cf16576ee597