Run browspr in the extension.
The browser extension contains wspr in a nacl plugin.  For now this
plugin is checked into the repo in extension/nacl/main-64.nexe.  Once I
add the veyron-go-compiler to environment repo, building the nacl plugin
can happen as a build step.

If you have veyron-go-compiler installed and want to build the plugin
yourself, you can do

  $ NACLGOROOT=/path/to/veyron-go-compiler cd extension && make nacl

The browser-integration tests (make test-integration-browser) build the
extension into extension/build-test, and run the tests in a browser with
that extension loaded.  We use a seperate build directory for the test
extension because it has the mounttable and proxy addresses set via
environment variables, and we don't want to risk building that version
into the normal extension/build directory.

NOTE: While the tests pass, this version of the extension isn't usable
for apps yet because I have not fixed the auth flow.  That is next on my
list.

There are also a lot of minor test fixes, because there are many places
where the tests hang or crash if an error occurs.

Change-Id: I33b0580267b040ad894a916fe65d28ab32cad7e6