Better error message when veyron.js can't connect to wspr.
We were logging a helpful message before, but since the default log level is
NOLOG (!?) it wasn't being displayed.  The error that we were returning
wasn't very useful: "connect ECONNREFUSED".

This change puts a helpful message right in the error that is returned:
"Failed to connect to wspr at url ws://localhost:8124/ws: connect
ECONNREFUSED"

Change-Id: Id421f2b32b14012281b20ba5b4585030ebc8b4ec