veyron.js/test_service/test_serviced: Convert to verror2.
This CL converts the code in veyron.js/test_service/test_serviced
to use verror2 instead of verror.
The import statements rename verror2 to verror to make the
anticipated future renaming of verror2 to verror simpler when it happens.

The changes to test/integration/test-service-errors.js are to
compensate for the changes in the error messages generated.
(verror allowed the develoepr to associated an arbitrary error message
with an error code, perhaps different on each use.  verror2 requires that
the (error-code,language) pair map to a given string, so the test must
check for the strings currently set up for US English.)

Beware that the test test_serviced_test.go is considered obsolete, and does not
compile or pass even before this change.

I reused the standard errors where is seemed appropriate.
This will change the details of the error messages associated with the errors.

Change-Id: Ib18a7fa93cf2e178d48d9332b7e7611376e81b17