Put timeout/connect logic in the event proxy.
The event proxies are the points of connection between the web app and
the content script.

This CL adds logic to the event proxies so that if the extension is not
running, the web app will be notified.

Similar logic had existed in the "auth" stuff inside veyron.js, but that
is no longer necessary.

Tests were added.

I gutted the "auth" tests because they were testing the same thing, and
I need to write real tests there.

Change-Id: Ie447d1e310d25917df78cdd51c97fbb4cf93b32f