Browspr Authentication support.
When a veyron web app asks for authentication, it sends an 'auth'
message to the extension, which creates and associates an account on
WSPR.

Previously, the extension was communicating with WSPR over HTTP.

Now, the extension must communicate with Browspr over postMessage.

This CL moves the old wspr HTTP logic to RPC channels.

The RPC channel implementation is mostly unchanged from Benj's original
implementation.  That is is extension/lib/background/channel.js

The authentication logic is encapsulated in
extension/lib/background/auth-handler.js

There were also some refactorings that had to happen in
extension/lib/background/index.js to make this work.

MultiPart: 1/2

Change-Id: I2594c77e0b9823eacd90da8f43c7b93ec4f686b1