veyron.js: Changed the Invoker to allow []InterfaceSig as the description.
We now allow a Service to implement multiple Interfaces and we allow
the developer to specify all the Interfaces that the Service matches.

This change also changes the logic in the Signature constructor so we
only include the methods in the Signature that match the Interface
passed in.  This allows us to constructor the proper signature response
when a service implements two Interfaces.

If there are any methods that are not accounted for in the Interfaces
passed in, we create a final Signature that has the left over methods.

Change-Id: I411361626eb9a06847e9018bfa87b9e8f836c8df