Refactors service reflection and invocation.
* Fixes veyron/release-issues#635
* Refactors Invoker#invoke and it's tests.
* Incorporates changes from https://vanadium-review.googlesource.com/#/c/1137
* There were several other issues that came out of this CL which should be
addressed separately: veyron/release-issues#657 veyron/release-issues#678

When defining a service invocations will always call a service method with a
Context object as the first argument, a callback as the last argument, and a
potential injectable $stream object.

The src/lib/arg-helper module was refactored into src/lib/arg-inspector to
prevent excessive looping. The ArguementInspector objects can be instantiated
and passed around for checking a service method's argument structure for
invocation and injections. There are still several optimizations that can be
done but this gets things moving in that direction.

Change-Id: Id019ef3ec19a563aa142337b89a82a0094942556