javascript: Fix flaky cancellation test.
This test was broken when we transitioned to requiring context as the
first parameter to server methods.  A helper method was incorrectly
turned into a server method which cased the real argument to be lost.

This only breaks sometimes because the method is only important if
the waitforstatus call arrives before the neverreturn call, and
that only happens half the time.

Change-Id: Iab8787921207e4c836df3904306228781f91c042