Use the same extension ID in production and tests.
We used to have a "dev" key, which caused the extension to have its own
ID.  There is no longer any reason to do this, as we can use the
production ID in the dev extension.

Doing this simplifies our environment, since the identity server no
longer needs to know about both extension IDs.

This CL also updates the README.

Change-Id: I0322bec9245d5a9a25bd67895129060cf580d10e