Treat errors during plugin startup as crashes.
Currently, if the plugin errors during start, we log an error in the
background page console, but don't do anything else with that error.

This CL makes errors during start-up get treated like plugin crashes,
which will notify all veyron apps.

The error message gets threaded through the messages, so the app can
display it and/or the developer can use it to debug.

Change-Id: Ia6f97ad70e7cdf0929c46fee7d398813d88ae04b