js.core: Rename things called *Context to *Call.
This is part of an effort to clarify naming around contexts.  We are
going to rename ServerContext and security.Context to ServerCall and
security.Call respectively.  Now when people say "context" we will
know they mean context.T.

In a follow-up change I can try to change the variable names ctx and
context to call.  I also need to merge the security.Call interface into
ClientCall and remove redundant access to LocalPrincipal.

MultiPart: 6/11

Change-Id: I586e0a17bfc0b74fd2869f2e7d7ea7b2c5f4d2f7