js/vdl: Add any(foo) => foo conversion
This helps us in the deepwrap => shallow wrap case.
(foo => any(foo) was already working)

Previously,
{any
  val: {string
    val: 'foo'
  }
}

would not convert to Types.STRING
because we only did a single unwrap (the top-level any)
Thus, { val: 'foo'} was not convertible to 'foo', since it's an object.

We needed to unwrap once more, and we now will do so under specific conditions.

Change-Id: I48fd1221285aa0f20a5ecc0095e75d147555ec4a