js/core: A first cut at the layout of the jsdoc output
This is a small change, so we can see if layouting the docs in this way
makes sense.  This goes down the path of having a top level vanadium
module with a bunch of namespaces that it holds.  Then in the places
where we have the methods/object exported, we tag them with memberof
tags to build the heirachy

Change-Id: I5542a1a016d9439935275fe1c010f2cccc5db822