js/core: Protect JS clients from bad input
Convert all output to what the signature said it ought to be.
Capture the error, if canonicalization fails.

Note: Doesn't protect outStream (if the client receives streaming output)
Then again, I'm not sure where outStream is since we combine Duplex.

Will modify proxy/stream.js to protect the read types.
This will be done in the same CL that adds typed stream tests and
tests that we're protected against lying streams.

Change-Id: I36a89fcd605d6d0bef34e10b27ba3e119a9b1798