Add ability to set namespaceRoots and proxy in config to vanadium.init().
This involved more plumbing from webapp to wspr, and a good amount of
refactoring along the way.

In the browser now, you must send the "create-instance" message before
you can send any browspr messages.  I think this is actually cleaner and
safer than the "lazy-init" we were doing before.

I'm not sure how to test the case where you pass in a proxy to
vanadium.init.  That is still a TODO.

MultiPart: 1/2

Change-Id: I2f1b42e221964ed7841013201aac60b29fda643f