js/core: Typed Streams
Previously, only the writing side of the Duplex was canonicalized.
This was mostly fine for tests, but the reading side was not protected.

Added a call to canonicalize on each item to be read ensures that the
data is converted to the correct type.

Tests were added for:
- inStream only (string)
- outStream only (number)
- non-native inStream and outStream ([]bool)
- differing types inStream(num) outStream(struct)
- any to any identity stream

Tests have not been added for:
- sending null (https://github.com/veyron/release-issues/issues/1332)
- lying streams (e.g., outStream type doesn't match signature)
  Will be added later

Change-Id: I01e426eb44804bba20cbd4b8c8f36791f9a6176d