Allow CallOptions to be passed to service methods.
This CL allows you to pass in rpc call options when making a vanadium
rpc.

Currently the only supported call option is 'allowedServersPolicy'.

Ex.

  client.bindTo('service/name', function(err, service) {
    var opts = client.callOpts({
      allowedServersPolicy: someBlessings
    });
    service.method(ctx, arg1, arg2, opts, callback);
  });

Technically, the call opts can be passed anywhere between the ctx and
the callback, but I think we should encourage them to be before the
callback, although this is not enforced.

MultiPart: 1/4

Change-Id: I0e6416f8d710e0f60c0d81b744b8a7ce0bb9b8d3