js/core: VanadiumError display some native types nicely
https://github.com/veyron/release-issues/issues/1560

We were previously guessing JSValue for the paramList.
Unlike signature, this is less acceptable since other languages are
allowed to receive our errors.

This CL guesses string, number, and bool for the user if they happen
to provide that to us. Note: This means they will be wrapped in
the paramList post-decode.

The problem is difficult. We want to allow users to construct errors
with native values without forcing them to specify the exact types they
want to send across the wire. Our guessing system will need to be quite
robust to satisfy these constraints.

Change-Id: Ibe40ad688afda09883baa7b1a1806df19296e0c0