naming: Utility function to extract blessing names from v4 endpoints.
Arguably, the better fix is to have a JavaScript equivalent of the Go
naming.Endpoint object (and v23.NewEndpoint factory function). However,
such a change is well beyond the current JavaScript expertise of the
author. So, the author requests this change for the related CL
(which will in turn un-block a change that removes large swaths
of dead code around MountedServer.BlessingPatterns) for now and will
seek guidance for the Endpoint type in JavaScript.

See also:
- https://vanadium-review.googlesource.com/8649
- https://vanadium-review.googlesource.com/#/c/8648/
- https://vanadium-review.googlesource.com/#/c/8650/

MultiPart: 3/4
Change-Id: Id4bce478b0fe690c90a15e2efd835963276b7d7e