js/core: Rework native types and canonicalize
Native types (especially error) have not interacted well with
canonicalize's expectations.

For now, this CL makes it act a bit more like JSValue.

The Native Type Registry was changed to take more type information when
converting a native value to wire value. The same had to be changed in
the Date and Error converting functions; they had to handle values that
were already in wire form.

Change-Id: I49e0ca7cc052258850c817b6e4ed59877f4fc3e5