extension: Add a styled caveat page to the extension oauth flow.
Changes:
(1) Users will be prompted before before granting a blessing to an origin.
(2) Ability to add caveats is reenabled.
(3) Users will be warned if the site asking for a blessing is not https.
(4) Users have a cancel button that cancels the blessing.

Screen shot of the page: https://screenshot.googleplex.com/rX4QSwoY3O.png.

Change-Id: Ib0d9026b9834914846aaa0a8c3224ab38aa81ca8