js/core: Use varargs in Error instead of []param
We occasionally used to create Errors like this
new verror.InternalError(ctx, [param1, param2, ...])

This led to our Error format string containing lots of '?' characters.

This wasn't working because we don't treat the 2nd param as
varargs when it's an array. Further, since VError is specifically
allowed to have any kind of param in the paramList, we can't
assume that any array must be treated like a vararg.

Thus, instead of modifying VError to construct the above way, this
CL switches to varargs.
new verror.InternalError(ctx, param1, param2, ...)

This also has the benefit of making the error messages correct again;
they weren't properly filling in the arguments the above way.

Tests added for the affected files:
- invoker
- client
- server-router

Change-Id: I2e8eb8608e1800c80e27e80a530402bcf760bbcf