js/core: Set publicKey in Blessings
We were accessing publicKey in a Blessings object, when it had _key.
_key => publicKey

This led to a few other minor changes.
- In test-server-context, we were returning the Context via a JSValue.
With _key as publicKey, it was being sent over instead of being treated
like a private field. We no longer check that since the publicKey isn't
known beforehand.
- AclAuthorizer was changed to use the new style of authorizer; throw
  when rejecting and return nothing to allow
- Integration tests added in test-authorizer.js

rebase (to force Jenkins to run)

Change-Id: I3cc4e9230bd6daaedfe04949ae24954ddd124c63