js/core: Share an encoder and decoder for all clients.
This improves RPC performance by 50% because the most expensive part
of making an rpc is encoding the type message.  This is really an
interim change, since the real solution is to have a separate type
flow that is shared by all messages from wspr to go.

This improves: https://github.com/vanadium/issues/issues/421

MultiPart: 1/3
Change-Id: I48ae5a43cc5d7b9e47a94bb8c677c2c37b3d126e