js/core: Add WebDriver library for UI Testing
This adds a library for testing the Vanadium extension with any of
the sub-projects that depend on JavaScript.

The library has code to install the Vanadium Extension but has no
tests on its own.

Use of this library will be optional since it needs many components
- Chrome WebDriver
- Google username/password
- Xvfb
- maven

This part of the change does the following
- Adds the library to test/ui
- Adds a README to explain how to use the library.

MultiPart: 2/4
Change-Id: Id07d04a3938fa4a6fb878ba67f50113d1805d241