js/core: Extension Caveat tab race fix
Fix race with the account name.
By "locking" early (not in a callback) and setting the account name
immediately, we ensure that any later getCaveats calls will reuse the
same outstanding request tab.

Removed unnecessary the authState flag and corresponding check.
Was this needed? It's a random value, so any (racing) tabs
won't know it's actual value. We also have requestId if we need things
to be unique.

Removed unnecessary _lastRequestId and requestId. The former was an
internal tracking variable that was unused. The latter was sent in
the url, but was also unused.

(Tested on HelloPeer using the build-dev extension.)

Change-Id: I2e9fd7dac3191c3e3c91b5e49aaf6d5cf24eaa42