js/core: Make decode async.
This is step 1 in a multi step process to use a separate type flow
between go and js.  The separate type flow will allow us to use a new
encoder/decoder for each rpc, but share the type messages between them.
This should result in a huge perfomance improvement.

I'm sending this change out separately because it's a monster and I
don't want to hide the real type flow change in this monster.  In the
process, I also found a bug in the caveat validator that failed to
validate async caveat chains when there were more than 1 caveat in the
chain.

Change-Id: Ida1814a1715e1bad492bcba4c4d522a576906ec7