TBR js/core: Give extension more time to prepare
This may alleviate the errors in the web tests for namespace browser
and www. Hypothesis: Accessing a page that needs to talk to the extension
too soon, leads to the page failing to initialize the caveat tab.

However, this merely covers up another problem.
https://github.com/vanadium/issues/issues/516

Give extension even more time to get ready

Change-Id: I690739590fbfc1b413a5284cfe02852f309d2419