js.core: Fix bug in vom zero-struct-field encoding.
This bug was reported by Adam Sadovsky.  The bug occurs with the
following example.

type X struct { A any }

The buggy code encodes X{false} the same as X{nil} - in general
any value of the A field that is the zero value for its type is
dropped.  That's just wrong.

The fix is in the encoder.canIgnoreField logic.

In addition I noticed that we'd inadvertantly exposed too many
methods from vom.Encoder.  I've fixed that as well.

Change-Id: Ibf1243acd16877f47765992d48480c06ac993a8f
MultiPart: 3/3