js/core:  Fix several cases where we were logging about unhandled stream close messages.
There is still a race here unfortunately, where the javascript server
response is sent while the stream close message is on the wire.  This
is annoying but the behavoir is still correct.  There were a few other
cases that were fixed where we weren't correctly handling cancels
for streaming rpcs.

This cl also includes a small bit of refactoring of handleRPCRequest.

Change-Id: I1fcdc538c61c711cda14b5e63cc3b27832c6e84c