javascript: Server and Client API Changes.
This with CL JavaScript' server API is changing
to match the new server API in Go and also client
API is chaning a bit so it is more consistent.

Server API Changes:
Changes are made to align with the server API
changes in Go that was submitted with http:/v.io/c/15257

In short, server.serve and server.serveDispatcher were
removed from the server object and new asynchronous methods
were added to runtime so users can create a server
object at the same time as starting the server.

`
var server = rt.newServer();
server.serve(name, serviceObject, cb);
`

is changing to:

`
rt.newServer(name, serviceObject, cb);
`

a server object will be passed to cb.

With this change, runtime.newServer no longer
immediately returns a server object, instead it also
takes in the serviceObject and a name and creates AND
starts the server before returning it.
new method runtime.newDispatchingServer was also added to
replace the server.serveDispatcher.

Client API Changes:
Now that runtime.newServer is asynchronous, having
runtime.newClient be synchronous would have been
inconsistent, also Go side is removing the ability
to create multiple clients anyway. Therefore with
this cl runtime.newClient was renamed to runtime.getClient
which would return a singleton client object.

runtime.namespace was also renamed to runtime.getNamespace
to be consistent with runtime.getContext and runtime.getClient

MultiPart: 2/10
Change-Id: I47415bfa42070f36c6b2f08c53ef5fb034cb320f