js.core: Introduce options.ConnectionTimeout and remove options.RetryTimeout.
There were a few subleties during implementation:
(1) We needed to introduce flow.Manager.DialCached to ensure that
ConnectionTimeout == 0 means "only use a cached conn".
(2) To implement flow.Manager.DialCached, we need to introduce
conncache.FindWithRoutingID to ensure that we only return
connection to end servers when a rid is present, instead of
a connection to an intermediate proxy which is what would
happen if we proceeded with the network lookup.
(3) When retrying connection establishment, ConnectionTimeout == 0
must also imply options.NoRetry, otherwise we will unnecessarily
retry a connection establishment that is not in the cache. This
is not true for verror.RetryBackoff errors (errors returned from
applications) since they may be requesting a retry for arbritrary
reasons.

Change-Id: I1a5acf5e3b0fdccf511724de394f283357db53fd
MultiPart: 3/3